firmware: fix compiler warnings

+++ firmware/common/LPC17xx_Startup.c
@@ -60,7 +60,7 @@ extern int main(void);
 /* Reset Handler */
 void Reset_Handler(void)
 {
-    unsigned long *src, *dest;
+	unsigned long *src, *dest;
 
 	// Copy the data segment initializers from flash to SRAM
 	src = &_etext;
@@ -76,11 +76,11 @@ void Reset_Handler(void)
 		*src++ = 0;
 	}
 
-    __libc_init_array();
-    
-    // Set the vector table location.
-    SCB_VTOR = &_interrupt_vector_table;
-    
+	__libc_init_array();
+
+	// Set the vector table location.
+	SCB_VTOR = (uint32_t)&_interrupt_vector_table;
+
 	main();
 
 	// In case main() fails, have something to breakpoint
